fix: fix all Dtos, add some fields, add ApiError entity

+++ mainApp/src/main/java/ru/practicum/ewm/location/Location.java
@@ -6,7 +6,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 @Data
 public class Location {
-    private Double lat;
+    private Float lat;
 
-    private Double lon;
+    private Float lon;
 }
